requirements.txt
faker
django
cloudinary
djangorestframework